chore: refactor admin contest folder structure
